add status report
