+++ types/copilotkit.d.ts
@@ -1,6 +1,9 @@
-declare module '@copilotkit/react-ui' {
+declare module '@copilotkit/react-core' {
   export const CopilotKit: any;
   export const useCopilotReadable: any;
+}
+
+declare module '@copilotkit/react-ui' {
   export const CopilotSidebar: any;
   export const CopilotChat: any;
 }
